requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
attrs==25.3.0
audioop-lts==0.2.1
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
cowsay==6.1
cycler==0.12.1
discord==2.3.2
discord.py==2.5.2
fonttools==4.59.2
frozenlist==1.5.0
idna==3.10
kiwisolver==1.4.9
matplotlib==3.10.6
multidict==6.1.0
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
propcache==0.3.0
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
pytz==2025.2
requests==2.32.5
seaborn==0.13.2
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
yarl==1.18.3
